JetBrains plugin: release 3.0.0-alpha.1
