--- client/jetbrains-shared/.gitignore
@@ -1,40 +1,0 @@
-# Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio and Webstorm
-# Reference: https://intellij-support.jetbrains.com/hc/en-us/articles/206544839
-
-*.jar
-!gradle-wrapper.jar
-
-# User local IDEA configuration files
-.idea/
-
-# IntelliJ project
-*.iml
-
-# Build output & caches for IntelliJ plugin development
-build/
-idea-sandbox/
-.gradle/
-
-## File-based project format:
-*.iws
-
-# IntelliJ
-/out/
-
-# mpeltonen/sbt-idea plugin
-.idea_modules/
-
-# JIRA plugin
-atlassian-ide-plugin.xml
-
-# Crashlytics plugin (for Android Studio and IntelliJ)
-com_crashlytics_export_strings.xml
-crashlytics.properties
-crashlytics-build.properties
-fabric.properties
-
-# JavaScript resources
-src/main/resources/dist/*
-
-# VSCode langserver artifacts
-/bin/
