--- client/jetbrains-shared/gradle.properties
@@ -1,9 +1,0 @@
-# Probably would be nicer if this was not duplicated across projects,
-# but not sure how to dedupe. It's fine for now.
-
-
-# Java language level used to compile sources and to generate the files for - Java 11 is required since 2020.3
-javaVersion=11
-# IntelliJ Platform Properties -> https://github.com/JetBrains/gradle-intellij-plugin#intellij-platform-properties
-platformType=IC
-platformVersion=2021.1
